Process: move definitions to cpp
